chore: update return type to Optional[str]
Updated the return type of the methods `get_provider_model_id` and
`get_llama_model` in the `ModelRegistryHelper` class to `Optional[str]`
to indicate that they may return a string or None when no match is
found. This change improves the clarity of the methods' behavior and
supports better type safety.

Signed-off-by: Sébastien Han <seb@redhat.com>